refactor: standardize filenames
